requirements.txt
torchsummary==1.5.1
numba==0.45.0
catboost==0.12.2
pandas==0.25.0
scipy==1.4.1
tensorflow==2.1.0
torch==1.1.0
matplotlib==3.0.3
tqdm==4.30.0
albumentations==0.4.3
efficientnet_pytorch==0.6.3
scikit_image==0.16.2
optuna==0.7.0
torchvision==0.3.0
seaborn==0.9.0
lightgbm==2.2.3
opencv_python==4.2.0.32
segmentation_models_pytorch==0.1.0
numpy==1.16.3
Pillow==7.1.2
category_encoders==2.2.2
japanize_matplotlib==1.0.5
matplotlib_venn==0.11.5
skimage==0.0
scikit_learn==0.22.2.post1
